Updating offset, deferred initialization (#1134)
* * Removed offset from round robin, dynamic_load and auto tune policies
* Made deferred initialization a struct and added constructors for that
* Updated the tests
* Switched offset in fixed resource policy to size_t
* Fixed dynamic selector comment in CMakeLists
* Changed runtime_error to logic_error

* Update fixed_resource_policy.h

* updating deferred init struct

* Update sycl_backend.h

* Update dynamic_load_policy.h

* removing offset from auto tune

* dynamic load test update

* removing template parameters from auto tune selection type

* updating offset tests for fixed policy

* Update test_fixed_resource_policy_inline.pass.cpp

+++ test/parallel_api/dynamic_selection/test_fixed_resource_policy_inline.pass.cpp
@@ -6,16 +6,15 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
-
 #include <iostream>
 #include "oneapi/dpl/dynamic_selection"
 #include "support/inline_backend.h"
-#include "support/test_dynamic_selection_utils.h"
+#include "support/test_offset_utils.h"
 
 int main() {
   using policy_t = oneapi::dpl::experimental::fixed_resource_policy<TestUtils::int_inline_backend_t>;
   std::vector<int> u{4, 5, 6, 7};
-  auto f = [u](int i, int offset=0) { return u[offset]; };
+  auto f = [u](size_t i, size_t offset=0) { return u[offset]; };
 
   constexpr bool just_call_submit = false;
   constexpr bool call_select_before_submit = true;
